Update patch set 1
Patch Set 1:

> If this is Beijing, would we use 1.2.0-SNAPSHOT?

I prefer to use the stable release version.  If changes are made to the logging library for Beijing, I think that we should be able to upgrade (or not upgrade) the various services which use the library independently depending on whether or not they require the changes.

As we've seen, SNAPSHOT artifacts are not retained indefinately in nexus.  My preference is to depend on released artifacts when possible, only upgrading to the latest snapshot in the event that a service needs to pick up a bug-fix/feature. 

If no changes go into the logging library for Beijing (or only changes that the search-service doesn't care about), I don't see any reason why the search service couldn't continue to use the 1.1.0 version of the library for Beijing and beyond.

Patch-set: 1
Reviewer: Gerrit User 453 <453@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0