Update patch set 1
Patch Set 1: Code-Review+1

I am a bit confused about the versioning mechanism.
Someone created 1.2.1 version and Arul changed the version to 1.2.2-SNAPSHOT.
So, 1.2.1 release is exactly the same as 1.2.0.

I would have expected that the version would remain 1.2.1-SNAPSHOT until Arul's change is pushed and that we decide to make his changes the 1.2.1 release (and then prepare for the next patch by changing the version to 1.2.2-SNAPSHOT)

Patch-set: 1
Label: Code-Review=+1