Update patch set 3
Patch Set 3: Code-Review+1

I'm surprised those are the only changes.  That's cool.  Once all our services are interworking with ES 6.x + latest Search service then it maybe worthwhile just making the changes in Search so that all the APIs are compliant to ES6 syntax and then we wouldn't need to translate the requests any more.

I don't think we have any specific requirements that Search has to interwork with different versions of ES, and I think the requirement to adapt to current ES 2.4.1 syntax was to simply to ease adoption.

Patch-set: 3
Label: Code-Review=+1